requirements.txt
djangorestframework~=3.14.0
django~=4.2